+++ .eslintignore
@@ -2,3 +2,4 @@
 /dist
 /node_modules
 /test/fixtures
+/test/outputs
