+++ .prettierignore
@@ -2,4 +2,5 @@
 /dist
 /node_modules
 /test/fixtures
+/test/outputs
 CHANGELOG.md
