+++ src/utils/ready.js
@@ -4,7 +4,7 @@ export default function ready(context, callback, req) {
     return callback(context.stats);
   }
 
-  const name = req.url || callback.name;
+  const name = (req && req.url) || callback.name;
 
   context.logger.info(`wait until bundle finished${name ? `: ${name}` : ''}`);
